see line 57, I had to hard code the hi_path
Could not use hiscore.ini, had to hard code it.  If I ever come back to this I might change the regex in the string match line to look for the two groups separated by one or more spaces, I think right now it is looking for exactly one space!